Fixes creation of an empty dir on output directory
During the first time a methylation analysis is run over a project,
an empty directory in the output directory is created with the name
of the current user. This is a problem related with two runs of Picard
tools (sorting and sequence .dict creation). We simply check if such
an empty directory exists and remove it.