Fixes bug of empty .methylation files at the end of methylation-analysis
A stream to the .methylation files was created at the start of the
analysis and not closed. In multi-threaded analysis the .methylation
is first created in UUID-random.methylation files that are merged
in treeReduce operations. At the traversal end, the last UUID-random
files are renamed to the final .methylation, but this "open" stream
impede this operation in some file systems (windows-docker, for example).
This lost stream is no longer required and thus removed.