Adds option for remove SAM files after converting to BAM
In order to save space, "-m" option was added to analyze-methylation
command. This command sorts and converts to BAM the SAM output of
the previous alignment step. With this option, SAM files are removed
just after being sorted, and sorted files are removed just after
being converted to BAM and indexed.